Separated pump_variable from pump_electricity class
