+++ watertap/unit_models/zero_order/pump_electricity_zo.py
@@ -90,70 +90,3 @@ class PumpElectricityZOData(ZeroOrderBaseData):
 
         self._perf_var_dict["Electricity (kW)"] = self.electricity
         self._perf_var_dict["Applied Pressure (bar)"] = self.applied_pressure
-
-
-@declare_process_block_class("PumpVariableEfficiencyZO")
-class PumpVariableEfficiencyZOData(PumpElectricityZOData):
-    """
-    Zero-Order model for a low pressure pump with variable efficiency.
-    Low pressure correlation: Kuritza, J. C.(2017). https://doi.org/10.1590/2318-0331.0217170018
-    """
-
-    def build(self):
-        super().build()
-        self.flow_bep = Var(
-            units=pyunits.m**3 / pyunits.s, doc="Best efficiency point flowrate"
-        )
-
-        self.flow_ratio = Var(
-            self.flowsheet().config.time,
-            units=pyunits.dimensionless,
-            bounds=(0, None),
-            doc="Ratio between instantaneous flowrate and best efficiency point flowrate",
-        )
-
-        self.eta_ratio = Var(
-            self.flowsheet().config.time,
-            units=pyunits.dimensionless,
-            bounds=(0, 1),
-            doc="Ratio between the true efficiency and the best efficiency due to variable operation",
-        )
-
-        @self.Constraint(
-            self.flowsheet().time, doc="Constraint for variable pump flowrate ratio"
-        )
-        def flow_ratio_constraint(b, t):
-            return b.flow_ratio[t] == b.properties[t].flow_vol / pyunits.convert(
-                b.properties.flow_bep, to_units=pyunits.m**3 / pyunits.s
-            )
-
-        @self.Constraint(
-            self.flowsheet().time, doc="Constraint for variable pump efficiency"
-        )
-        def eta_ratio_constraint(b, t):
-            if b.config.process_subtype is "variable_lowpressure":
-                # if the flow is too far from the design point, the efficiency is set to 40%
-                if b.flow_ratio[t] < 0.6 or b.flow_ratio[t] > 1.4:
-                    return b.eta_ratio[t] == 0.4
-                # otherwise, use the correlation
-                else:
-                    return b.eta_ratio[t] == (
-                        -0.995 * b.flow_ratio[t] ** 2 + 1.977 * b.flow_ratio[t] + 0.025
-                    )
-            else:
-                raise ValueError(
-                    "Correlation type ", self.config.correlation, " not recognized"
-                )
-
-        @self.Constraint(
-            self.flowsheet().time, doc="Constraint for actual pump energy consumption"
-        )
-        def electricity_consumption(b, t):
-            return b.electricity[t] == pyunits.convert(
-                b.lift_height
-                * b.properties[t].flow_vol
-                * b.properties[t].dens_mass
-                * Constants.acceleration_gravity
-                / (b.eta_pump * b.eta_motor * b.eta_ratio[t]),
-                to_units=pyunits.kW,
-            )
